further refactor
now compiles using msvc
