+++ src/fmi4cpp/fmi2/import/fmi2CoSimulationLibrary.cpp
@@ -25,6 +25,7 @@
 #include "FmiLibraryHelper.hpp"
 #include <fmi4cpp/fmi2/import/fmi2CoSimulationLibrary.hpp>
 
+using namespace fmi4cpp;
 using namespace fmi4cpp::fmi2;
 
 fmi2CoSimulationLibrary::fmi2CoSimulationLibrary(const std::string &modelIdentifier,
